requirements.txt
google-genai>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
cryptography>=43.0.0
transformers>=4.45.0
torch>=2.1.0
accelerate>=0.30.0
soundfile
librosa
